Pull vector unit into separate source/header
