+++ riscv/riscv.mk.in
@@ -42,6 +42,7 @@ riscv_hdrs = \
 	jtag_dtm.h \
 	csrs.h \
 	triggers.h \
+	vector_unit.h
 
 riscv_install_hdrs = \
 	abstract_device.h \
@@ -94,6 +95,7 @@ riscv_srcs = \
 	jtag_dtm.cc \
 	csrs.cc \
 	triggers.cc \
+	vector_unit.cc \
 	$(riscv_gen_srcs) \
 
 riscv_test_srcs =
